+++ connectlib/__init__.py
@@ -1,4 +1,9 @@
+import logging
+
 from connectlib.__version__ import __version__  # noqa
 from connectlib.experiment import execute_experiment
+from connectlib.logger import set_logging_level
 
-__all__ = ["execute_experiment"]
+set_logging_level(loglevel=logging.INFO)
+
+__all__ = ["execute_experiment", "set_logging_level"]